Update README.md
Updated readme to say that this is the monorepo for tm35 metronome

Reason:
What became clear as the amount of repos grew was the fact that
if one of my libs change, or if Zig changes, then I'd have to update
all the repos. This became a large development overhead. This
mono repo will make development easier, but I'll have to figure out
how I do versioning now.