Rename "filter" to "command"
A command is just anything that takes our custom format and modifies it
in some way. It can filter, transform or add the format it doesn't
really matter.

The randomizer will should with a set of default commands which will
be split into different folders. Currently we only have the
"randomizers" folder.