Refactor tm35-load/apply to access the format through abstraction
We now have a `format.zig` which contains the datastructures that
are passed between programs encoded. tm35-load/apply now does not
outputer or parse strings directly, but instead maps the data to the
common datastructures and lets the abstractions handle parsing and
writing. This can also be done for all the other tools, but have not
been done yet.